+++ assets/src/apps/delivery/formats/adaptive/ActivityRenderer.tsx
@@ -6,7 +6,6 @@ import React from 'react';
 // the adaptive-activity web component should render
 // the parts
 const ActivityRenderer: React.FC<any> = (props: any) => {
-  // console.log('AR', { props });
   return (
     <div>
       {props.parts.map((partDefinition: any) => {