+++ assets/src/components/misc/resizer/types.ts
@@ -1,0 +1,9 @@
+export type Position = 'nw' | 'n' | 'ne' | 'w' | 'e' | 'sw' | 's' | 'se';
+export interface Point {
+  top: number;
+  left: number;
+}
+export interface BoundingRect extends Point {
+  width: number;
+  height: number;
+}