+++ assets/src/components/resource/editors/dragndrop/handlers/focus.ts
@@ -6,15 +6,16 @@ import { getFriendlyName } from '../utils';
 
 export const focusHandler = (
   setAssistive: (s: string) => void,
-  content: Immutable.List<ResourceContent>,
+  content: Immutable.OrderedMap<string, ResourceContent>,
   editorMap: ActivityEditorMap,
   activities: Immutable.Map<string, Activity>,
-) => (index: number) => {
-  const item = content.get(index) as ResourceContent;
-  const desc = item.type === 'content'
-    ? getContentDescription(item)
-    : getFriendlyName(item, editorMap, activities);
+) => (key: string) => {
+  const item = content.get(key) as ResourceContent;
+  const desc =
+    item.type === 'content'
+      ? getContentDescription(item)
+      : getFriendlyName(item, editorMap, activities);
 
-  setAssistive(
-    `Listbox. ${index + 1} of ${content.size}. ${desc}.`);
+  const index = content.keySeq().findIndex((k) => k === key);
+  setAssistive(`Listbox. ${index + 1} of ${content.size}. ${desc}.`);
 };